cambios scorecard
divisione ntre mil de refacciones

+++ HDBackend/HD_Ventas/Modelos/mdlCarga_Scorecard_porVendedor_Dash_TablaAsesor.cs
@@ -52,9 +52,9 @@
         public int Real_Visitas { get; set; }
         public int Comp_Visitas { get; set; }
 
-        public int Objetivo_Refacciones { get; set; }
-        public int Real_Refacciones { get; set; }
-        public int Comp_Refacciones { get; set; }
+        public double Objetivo_Refacciones { get; set; }
+        public double Real_Refacciones { get; set; }
+        public double Comp_Refacciones { get; set; }
 
 
     }
